+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Release date `UNRELEASED`
 
+- (refactor BREAKING): use hsts and referrer-policy from web-security
+
 ## `1.1.0-103-beta2`
 
 Release date `2023-03-20`